feat: compile server files to ts
