+++ src/server.ts
@@ -1,0 +1,6 @@
+import config from './common/config';
+import app from './app';
+
+app.listen(config.PORT, () =>
+  console.log(`App is running on http://localhost:${config.PORT}`)
+);
